Bug Fix Additional logic

+++ Pong/src/main/java/ch/teko/loefflee/BallDoubleSpeed.java
@@ -2,6 +2,6 @@ package src.main.java.ch.teko.loefflee;
 
 public class BallDoubleSpeed {
     public static void applyEffect(Ball ball) {
-        ball.setSpeed(ball.getSpeed() * 2);
+        ball.setSpeed((int) (ball.getSpeed() * 2));
     }
 }
